requirements.txt
streamlit==1.32.2
google-generativeai==0.3.2
langchain==0.1.14
tiktoken==0.6.0
sentence-transformers==2.2.2
rank_bm25==0.2.2
beautifulsoup4==4.12.3
pandas==2.2.2
unstructured==0.10.25
PyMuPDF==1.23.26
python-dotenv==1.0.1
torch==2.2.1
scikit-learn
